[#398] Refactor scenario modeling component intp smaller component
